requirements.txt
python-dotenv==1.0.1
requests==2.31.0
tabulate==0.9.0
urllib3==2.2.0
